[Fix] Fix the bug in binary cross entropy loss
 Fix the bug in binary cross entropy loss when using multi-label datasets e.g.VOC2007

+++ mmpretrain/models/heads/multi_label_cls_head.py
@@ -93,7 +93,7 @@ class MultiLabelClsHead(BaseModule):
         num_classes = cls_score.size()[-1]
         # Unpack data samples and pack targets
         if 'gt_score' in data_samples[0]:
-            target = torch.stack([i.gt_score for i in data_samples])
+            target = torch.stack([i.gt_score.float() for i in data_samples])
         else:
             target = torch.stack([
                 label_to_onehot(i.gt_label, num_classes) for i in data_samples
